fix(*): support aarch64
the call conventions of x64: with the large return value(>16B), a hidden pointer is used to save the return value, and it's in rdi, then the arguments are stored in rsi, rdx, rcx, r8, r9. with void returning, the arguments are store in rdi, rsi, rdx, rcx, r8, r9. On Aarch64, however, the result value is store in x8, then arguments are in x0-x7, the same order with void returning.